pyproject.toml
[project]
name = "pmf-tsfm"
version = "0.1.0"
description = "Process Mining Forecasting with Time Series Foundation Models"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Yongbo Yu"}
]
keywords = ["time-series", "forecasting", "foundation-models", "process-mining"]

dependencies = [
    # Core
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
    
    # Configuration
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    
    # Data
    "pyarrow>=14.0.0",
    
    # Logging
    "rich>=13.0.0",
    
    # Models - Chronos
    "chronos-forecasting>=1.4.0",
    
    # Models - Moirai (uni2ts)
    "uni2ts>=1.0.0",
    "gluonts>=0.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
pmf-inference = "pmf_tsfm.inference:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pmf_tsfm"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[dependency-groups]
dev = [
    "black>=25.12.0",
    "ruff>=0.14.8",
]
